💄 Design: 활동 게시물 툴바 추가 & 파일명 변경
